Add files via upload
An application that computerizes the conventional car sale procedure which we are aware of. This is a Java/HTML based application. This helps in managing data related to buyers and sellers of the cars. Business reports can also be generated and viewed.
This program should handle a new and used vehicles inventory with minimum 10 cars. Every detail of a vehicle like manufacturer name, date of purchase, Kilo meters ran( in case of used vehicles), price, make/model, car description picture/s of the car